Use webtest instead of Selenium for identique view tests (frontend not relevant here)
